refactor new logic
