+++ src/Filters/Filterable.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Seredenko\Filters;
+
+/**
+ * Interface Filterable
+ * @package Seredenko\Filters
+ */
+interface Filterable
+{
+    public function filter();
+}
