+++ src/Filters/OrLogicFilter.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Seredenko\LogicFilters;
+namespace Seredenko\Filters;
 
 use Seredenko\Operator;
 
 /**
  * Class OrLogicFilter
- * @package Seredenko\LogicFilters
+ * @package Seredenko\Filters
  */
 class OrLogicFilter extends BaseLogicFilter implements Filterable
 {
