--- src/LogicFilters/Filterable.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Seredenko\LogicFilters;
-
-
-interface Filterable
-{
-
-}
